backend/requirements.txt
Flask==3.0.0
Flask-Bcrypt==1.0.1
Flask-JWT-Extended==4.5.3
Flask-Cors==4.0.0
python-dotenv==1.0.1
requests==2.28.1
supabase==0.0.1